:recycle: user 컨트롤러 파일 이동 및 nickname 의존 제거로 인한 수정
트랜잭션 삭제,.,.,. 추후 추가 예정,, 흑흑
